requirements.txt
streamlit
requests
PyPDF2
pytesseract
Pillow
langchain
langchain_community
langchain_ollama
faiss-cpu
scipy
FlagEmbedding
openai
ffmpeg-python
mammoth
python-pptx
textract
python-docx
openai-whisper
